Added tests and stylesheet for UserInfo.js. Updated bootstrap and react-boostrap to latest versions. Removed unwanted tests for skills page, languages page and general info page.
