+++ package.json
@@ -10,7 +10,7 @@
     "formik": "^2.2.9",
     "react": "^17.0.2",
     "react-addons-test-utils": "^15.6.2",
-    "react-bootstrap": "^2.2.2",
+    "react-bootstrap": "^2.3.0",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
