+++ src/Pages/GeneralInfo.js
@@ -49,7 +49,7 @@ function GeneralInfo(){
                 touched,
                 errors, 
             }) => (
-                <Container>
+                <Container data-testid="general-info">
                     <Form noValidate onSubmit={handleSubmit}>                 
                         <Row className="mb-3">
                             <Col xs={12} sm={12} md={4} className="mb-sm-4 text-center">
