+++ src/Pages/SkillsPage.test.js
@@ -1,28 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import NavHeader from '../Components/NavHeader/NavHeader';
 import SkillsPage from './SkillsPage';
 import { MemoryRouter } from 'react-router-dom';
 import SkillRow from '../Components//Rows/SkillRow';
 
-test('NavHeader.js renders NavHeader', () => {
-    render(<MemoryRouter>
-        <NavHeader />
-    </MemoryRouter>);
-    const testNavbarText = screen.getByTestId('navbar');
-    expect(testNavbarText).toBeInTheDocument();
-});
-
-// Sign-out icon appears when isLogoutEnabled is true
-describe('NavHeader.js', () => {test('Sign-out icon appears when isLogoutEnabled is true', () => {
-    render(
-        <MemoryRouter>
-            <NavHeader isLogoutEnabled={true} />
-        </MemoryRouter>);
-    const signOut = screen.getByAltText('LogOut');
-    expect(signOut).toBeInTheDocument();
-});
-});
 
 // <SkillRows /> components render to the page
 describe('SkillRow.js', () => {test('<SkillRows /> components render to the page', () => {
@@ -35,23 +16,13 @@ describe('SkillRow.js', () => {test('<SkillRows /> components render to the page
 });
 
 // User should see a heading that reads List and Rank your Work Related Skills
-describe('SkillsPage.js', () => {test("User should see a heading that reads 'List and Rank your Work Related Skills'", () => {
+describe('SkillsPage.js', () => {test("User should see a heading that reads 'Soft Skills and Proficiency'", () => {
     render(<MemoryRouter>
         <SkillsPage />
     </MemoryRouter>);
-    const skillsHeading = screen.getByText(/List and Rank your Work Related Skills/i);
+    const skillsHeading = screen.getByText(/Soft Skills and Proficiency/i);
     expect(skillsHeading).toBeInTheDocument();
 });
 });
 
-// User should see a heading that reads 'Personal Information'
-describe('SkillsPage.js', () => {test("User should see a heading that reads 'Personal Information'", () => {
-    render(
-        <MemoryRouter>
-            <SkillsPage />
-        </MemoryRouter>);
 
-    const personalInfo = screen.getByText(/Personal Information/i);
-    expect(personalInfo).toBeInTheDocument();
-});
-});
