+++ src/Pages/UserInfo.css
@@ -1,0 +1,27 @@
+
+.accordion-button{
+    font-size: 1em;    
+}  
+
+.accordion-item {
+    border: 1px solid black;
+}
+.accordion-button:not(.collapsed) {
+    color: inherit;
+    background-color: #dc3545; 
+    color: #f9fafb;   
+} 
+
+.accordion-button:not(.collapsed)::after {
+    filter: brightness(0%) invert(100%);
+}
+
+.accordion-button:focus{
+    box-shadow: inherit;
+}
+
+
+
+
+
+
